# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kōri",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout-all": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout All",
        "operationId": "logout_all_api_auth_logout_all_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change Password",
        "operationId": "change_password_api_auth_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccountStatus": {
        "type": "string",
        "enum": [
          "new",
          "open",
          "closed",
          "other"
        ]
      },
      "AccountType": {
        "type": "string",
        "enum": [
          "broker",
          "iis",
          "invest_box",
          "invest_fund",
          "other"
        ]
      },
      "AnswerStyle": {
        "type": "string",
        "enum": [
          "concise",
          "detailed"
        ]
      },
...
      "BrokerAccountOut": {
        "properties": {
          "alias": {
            "type": "string",
            "pattern": "^acc[0-9]+$",
            "title": "Alias"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/AccountType"
          },
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "opened_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened At"
          },
          "is_hidden": {
            "type": "boolean",
            "title": "Is Hidden"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "alias",
          "name",
          "type",
          "status",
          "opened_at",
          "is_hidden"
        ],
        "title": "BrokerAccountOut"
      },
      "BrokerConnectionOut": {
        "properties": {
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "active",
                  "invalid"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "token_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Hint"
          },
          "last_verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At"
          },
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/BrokerAccountOut"
            },
            "type": "array",
            "title": "Accounts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "connected",
          "status",
          "token_hint",
          "last_verified_at",
          "accounts"
        ],
        "title": "BrokerConnectionOut"
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "unauthorized",
          "invalid_credentials",
          "forbidden",
          "registration_closed",
          "not_found",
          "conflict",
          "run_active",
          "email_taken",
          "broker_not_connected",
          "gone",
          "validation_error",
          "invite_required",
          "invite_invalid",
          "token_invalid",
          "token_not_read_only",
          "rate_limited",
          "user_budget_exhausted",
          "llm_busy",
          "llm_quota_exhausted",
          "tinvest_unavailable",
          "tinvest_rate_limited",
          "web_unavailable",
          "web_credits_exhausted",
          "disclosure_unavailable",
          "internal"
        ]
      },
      "ErrorOut": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "details",
          "request_id"
        ],
        "title": "ErrorOut"
      },
...
      "JsonValue": {},
...
      "LoginIn": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginIn"
      },
      "MeOut": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "settings": {
            "$ref": "#/components/schemas/SettingsOut"
          },
          "broker": {
            "$ref": "#/components/schemas/BrokerConnectionOut"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user",
          "settings",
          "broker"
        ],
        "title": "MeOut"
      },
...
      "ModelMode": {
        "type": "string",
        "enum": [
          "auto",
          "lite",
          "pro",
          "max",
          "ultra"
        ]
      },
...
      "PasswordChangeIn": {
        "properties": {
          "current_password": {
            "type": "string",
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "minLength": 10,
            "title": "New Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChangeIn"
      },
...
      "RegisterIn": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 10,
            "title": "Password"
          },
          "invite_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invite Code"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterIn"
      },
...
      "SettingsOut": {
        "properties": {
          "model_mode": {
            "$ref": "#/components/schemas/ModelMode"
          },
          "answer_style": {
            "$ref": "#/components/schemas/AnswerStyle"
          },
          "default_account": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Account"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model_mode",
          "answer_style",
          "default_account"
        ],
        "title": "SettingsOut"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "owner"
            ],
            "title": "Role"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "role",
          "created_at"
        ],
        "title": "UserOut"
      },
...
    }
  }
}